Create Item.cs
Example item script. Originally, I was planning on making a class for each item that derived from 'Item', and then override Use(). But I later decided to condense everything into one script.
Bug(s) discovered: none yet
Note that there are some additions, including:
isTeleport - whether to teleport the player or not
shieldUse - whether the shield is in effect or not
currentHealth - the current health of the player or the enemy
"Teleport Item", "Shield Item", "Health Item" - various items